+++ backend/leetcode_service/requirements.txt
@@ -1,0 +1,4 @@
+graphql_query
+gql[aiohttp]
+confluent-kafka==2.1.0
+pydantic==1.10.4
