+++ backend/leetcode_service/run-dev.sh
@@ -1,0 +1,5 @@
+set -o allexport
+source ../../docker-compose.env
+set +o allexport
+
+python main.py
