+++ backend/question_service/Dockerfile
@@ -1,0 +1,10 @@
+FROM python:3.10
+
+WORKDIR /backend/question_service
+
+COPY ./requirements.txt /backend/question_service/requirements.txt
+
+RUN pip install -r requirements.txt
+
+COPY . /backend/question_service
+COPY ./run.sh /backend/question_service/run.sh
