+++ backend/question_service/leetcode_service.py
@@ -1,0 +1,26 @@
+from aiokafka import AIOKafkaConsumer
+from config import get_config
+import json
+from database.question_collection import check_exist_question, create_question
+from database.mongodb import get_database
+from model.question import Question
+
+config = get_config()
+question_bank_consumer = AIOKafkaConsumer(config.kafka_topic_question_bank, bootstrap_servers=config.kafka_server_name)
+
+
+async def question_bank_consume():
+    try:
+        # consume messages
+        async for msg in question_bank_consumer:
+            question = json.loads(msg.value.decode('utf-8'))
+            db = await get_database()
+            exist = await check_exist_question(db, question["title"])
+            if not exist:
+                q = Question(**question)
+                await create_question(db, q.dict())
+    except Exception as e:
+        print(e)
+    finally:
+        # will leave consumer group; perform autocommit if enabled
+        await question_bank_consumer.stop()
