+++ backend/question_service/requirements.txt
@@ -1,0 +1,11 @@
+fastapi==0.85.0
+confluent-kafka==2.1.0
+pydantic==1.10.4
+pytest==7.2.0
+uvicorn[standard]==0.18.3
+graphql_query
+gql[aiohttp]
+motor==3.0.0
+aiokafka
+
+
