+++ backend/question_service/run.sh
@@ -1,0 +1,5 @@
+set -o allexport
+source ../../docker-compose.env
+set +o allexport
+
+uvicorn main:app --reload --port 5001 --host localhost
