--- backend/run.sh
@@ -1,2 +1,0 @@
-cd app
-uvicorn main:app --reload --port 5000 --host 0.0.0.0
