+++ start.sh
@@ -1,1 +1,6 @@
+
+set -o allexport
+source docker-compose.env
+set +o allexport
+
 docker-compose -f docker-compose.yml up
